backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
openai>=2.3.0
tiktoken==0.12.0
gunicorn==21.2.0
pydantic>=2.10.0
redis==5.0.1
marshmallow==3.20.1
psycopg2-binary==2.9.9
pg8000==1.30.3
cloud-sql-python-connector[pg8000]==1.5.0
